# Make Integer#zero? a separated method and builtin