--- a/.document
+++ b/.document
@@ -14,6 +14,7 @@ array.rb
 ast.rb
 dir.rb
 gc.rb
+integer.rb
 io.rb
 kernel.rb
 pack.rb