--- a/inits.c
+++ b/inits.c
@@ -82,6 +82,7 @@ rb_call_builtin_inits(void)
 {
 #define BUILTIN(n) CALL(builtin_##n)
     BUILTIN(gc);
+    BUILTIN(integer);
     BUILTIN(io);
     BUILTIN(dir);
     BUILTIN(ast);