--- a/integer.rb
+++ b/integer.rb
@@ -1,0 +1,9 @@
+class Integer
+  # call-seq:
+  #    int.zero? -> true or false
+  #
+  # Returns +true+ if +num+ has a zero value.
+  def zero?
+    Primitive.cexpr! 'int_zero_p(self);'
+  end
+end