A prerequisite to fix https://bugs.ruby-lang.org/issues/15589.
This commit alone doesn't make a significant difference:

```
$ benchmark-driver -v --rbenv 'before;after' benchmark/num_zero_p.yml --repeat-count=4
before: ruby 2.8.0dev (2020-06-14T22:01:15Z master 80bd3aa32b) [x86_64-linux]
after: ruby 2.8.0dev (2020-06-14T22:24:10Z builtin-zero 2a0a87df9a) [x86_64-linux]
Calculating -------------------------------------
                         before       after
             0.zero?   110.377M    109.280M i/s -     50.000M times in 0.452994s 0.457541s
             1.zero?    93.696M    109.119M i/s -     50.000M times in 0.533641s 0.458215s
            0r.zero?    19.031M     19.052M i/s -     50.000M times in 2.627243s 2.624463s
            1r.zero?    18.356M     18.533M i/s -     50.000M times in 2.723845s 2.697855s
            0i.zero?    16.849M     16.942M i/s -     50.000M times in 2.967477s 2.951229s
            1i.zero?    17.199M     17.044M i/s -     50.000M times in 2.907172s 2.933592s

Comparison:
                          0.zero?
              before: 110376618.2 i/s
               after: 109279938.5 i/s - 1.01x  slower

                          1.zero?
               after: 109119200.5 i/s
              before:  93695909.8 i/s - 1.16x  slower

                         0r.zero?
               after:  19051519.0 i/s
              before:  19031360.2 i/s - 1.00x  slower

                         1r.zero?
               after:  18533241.8 i/s
              before:  18356404.9 i/s - 1.01x  slower

                         0i.zero?
               after:  16942093.6 i/s
              before:  16849328.2 i/s - 1.01x  slower

                         1i.zero?
              before:  17198843.4 i/s
               after:  17043953.1 i/s - 1.01x  slower
```

This is subject to https://bugs.ruby-lang.org/issues/16961 discussion.